final commit

+++ Imari_Davis_P1/src/main/java/com/revatureimari/controllers/UserController.java
@@ -1,12 +1,12 @@
 package com.revatureimari.controllers;
 
+import com.revatureimari.Driver;
 import com.revatureimari.models.User;
 import com.revatureimari.services.UserService;
 import io.javalin.http.Handler;
 
 public class UserController {
     private UserService userService;
-    private User currentUser;
 
     public UserController() {
         userService = new UserService();
@@ -17,14 +17,18 @@ public class UserController {
     }
 
     public Handler createNewUser = context -> {
-        User user = context.bodyAsClass(User.class);
-        int userId = userService.createUser(user);
+        if (Driver.getCurrentUser() == null) {
+            User user = context.bodyAsClass(User.class);
+            int userId = userService.createUser(user);
 
-        if (userId > 0) {
-            user.setUserId(userId);
-            context.json(user).status(200);
+            if (userId > 0) {
+                user.setUserId(userId);
+                context.json(user).status(200);
+            } else {
+                context.result("Could not create user").status(400);
+            }
         } else {
-            context.result("Could not create user").status(400);
+            context.result("Current user must be logged out").status(400);
         }
     };
 
@@ -54,20 +58,29 @@ public class UserController {
         User user = context.bodyAsClass(User.class);
         user = userService.updateUser(user);
 
-        if (user != null) {
-            context.json(user).status(202);
+        if (Driver.getCurrentUser() != null && Driver.getCurrentUser().getUserId() == user.getUserId()) {
+            if (user != null) {
+                context.json(user).status(202);
+            } else {
+                context.result("Could not update user").status(400);
+            }
         } else {
-            context.result("Could not update user").status(400);
+            context.result("Invalid user. User data can only be modified by designated user.").status(400);
         }
     };
 
     public Handler deleteUser = context -> {
         User user = context.bodyAsClass(User.class);
 
-        if (user != null) {
-            context.status(200).json(userService.deleteUser(user));
+        if (Driver.getCurrentUser() != null && Driver.getCurrentUser().getUserId() == user.getUserId()) {
+            if (user != null) {
+                context.status(200).json(userService.deleteUser(user));
+                Driver.setCurrentUser(null);
+            } else {
+                context.status(400).result("Could not delete user");
+            }
         } else {
-            context.status(400).result("Could not delete user");
+            context.result("Invalid user. User data can only be modified by designated user.").status(400);
         }
     };
 
@@ -77,7 +90,11 @@ public class UserController {
         try {
             int userId = Integer.parseInt(param);
 
-            context.json(userService.deleteUserById(userId)).status(202);
+            if (Driver.getCurrentUser() != null && Driver.getCurrentUser().getUserId() == userId) {
+                context.json(userService.deleteUserById(userId)).status(202);
+            } else {
+                context.result("Invalid user. User data can only be modified by designated user.").status(400);
+            }
 
         } catch (NumberFormatException numberFormatException) {
             System.out.println(numberFormatException.getMessage());
@@ -88,12 +105,17 @@ public class UserController {
         User user = context.bodyAsClass(User.class);
 
         try {
-            currentUser = userService.loginUser(user);
+            if (Driver.getCurrentUser() == null) {
+                Driver.setCurrentUser(userService.loginUser(user));
 
-            if (currentUser != null) {
-                context.result("Logged in as " + currentUser.getFirstName() + " " + currentUser.getLastName());
+                if (Driver.getCurrentUser() != null) {
+                    context.result("Logged in as " + Driver.getCurrentUser().getFirstName() + " " +
+                            Driver.getCurrentUser().getLastName()).status(200);
+                } else {
+                    context.result("Invalid username or password").status(400);
+                }
             } else {
-                context.result("Invalid username or password").status(400);
+                context.result("Current user must be logged out").status(400);
             }
 
         } catch (NumberFormatException nfmException) {
@@ -106,21 +128,30 @@ public class UserController {
         User user = context.bodyAsClass(User.class);
 
         try {
-            currentUser = userService.loginUserByEmail(user);
+            if (Driver.getCurrentUser() == null) {
+                Driver.setCurrentUser(userService.loginUserByEmail(user));
 
-            if (currentUser != null) {
-                context.result("Logged in as " + currentUser.getFirstName() + " " + currentUser.getLastName());
-
+                if (Driver.getCurrentUser() != null) {
+                    context.result("Logged in as " + Driver.getCurrentUser().getFirstName() + " " +
+                            Driver.getCurrentUser().getLastName()).status(200);
+                } else {
+                    context.result("Invalid email address or password").status(400);
+                }
             } else {
-                context.result("Invalid email address or password").status(400);
+                context.result("Current user must be logged out").status(400);
             }
 
         } catch (NumberFormatException nfmException) {
             System.out.println(nfmException.getMessage());
-
         }
-
     };
 
-
+    public Handler logoutUser = context -> {
+        if (Driver.getCurrentUser() != null) {
+            Driver.setCurrentUser(null);
+            context.result("Successfully logged out").status(200);
+        } else {
+            context.result("No user is logged in").status(400);
+        }
+    };
 }
